agrync_backend/requirements.txt
asyncua==1.1.6
beanie==1.29.0
fastapi==0.115.11
motor==3.7.0
psutil==7.0.0
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.0.1
bcrypt==4.3.0
python-jose==3.4.0
requests==2.32.3
starlette==0.46.0
email_validator==2.2.0
python-multipart==0.0.20
uvicorn[standard]==0.34.0
concurrent-log-handler==0.9.25
pymodbus==3.9.1